pyproject.toml
[tool.poetry]
name = "systrader"
version = "0.1.0"
description = ""
authors = ["Kehinde Olalusi <kehindeolalusi75@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.1"
yfinance = "^0.2.40"
empyrical-reloaded = "^0.5.10"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
ruff = "^0.4.8"
pyfolio-reloaded = "^0.9.7"
pre-commit = "^3.7.1"
ipykernel = "^6.29.5"
alpaca-py = "^0.39.1"


[tool.mypy]
pretty = true
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
disable_error_code = "var-annotated, misc"
allow_redefinition = true
disallow_any_expr = false
exclude = ['tests']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
